refactor: refactoring handlers with OOP way.
using OOP to rewrite handlers with modularize pattern that would be easy maintenance and add more function in future.
Each handler extends a BaseToolHandler for consistent error handling and argument validation.
